@@ -11,13 +11,13 @@ namespace UserAuthenticator.Services
     {
         public Token CreateToken(User user, string role)
         {
-            Claim[] userRoles = new Claim[]
-            {
+            Claim[] userRoles =
+            [
                 new Claim("UserName", user.UserName),
                 new Claim("id", user.Id.ToString()),
                 new Claim(ClaimTypes.Role, role),
                 new Claim(ClaimTypes.DateOfBirth, user.Birthday.ToString())
-            };
+            ];
 
             var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes("8sfas9fg25as9f5a9sf5asfgaqsfuasfasfa"));
 
